third_party/WebKit/LayoutTests/virtual/deferred/inspector/timeline/timeline-decode-resize-expected.txt

Tests the Timeline API instrumentation of a DecodeImage and ResizeImage events


DecodeImage Properties:
{
    children : <object>
    counters : <object>
    data : {
        elementId : <number>
        imageType : "BMP"
        url : .../inspector/timeline/resources/test.bmp
    }
    endTime : <number>
    startTime : <number>
    type : "DecodeImage"
    usedHeapSize : <number>
}
DecodeImage Properties:
{
    children : <object>
    counters : <object>
    data : {
        elementId : <number>
        imageType : "GIF"
        url : .../inspector/timeline/resources/test.gif
    }
    endTime : <number>
    startTime : <number>
    type : "DecodeImage"
    usedHeapSize : <number>
}
DecodeImage Properties:
{
    children : <object>
    counters : <object>
    data : {
        elementId : <number>
        imageType : "ICO"
        url : .../inspector/timeline/resources/test.ico
    }
    endTime : <number>
    startTime : <number>
    type : "DecodeImage"
    usedHeapSize : <number>
}
ResizeImage Properties:
{
    children : <object>
    counters : <object>
    data : {
        cached : true
        elementId : <number>
        url : .../inspector/timeline/resources/test.ico
    }
    endTime : <number>
    startTime : <number>
    type : "ResizeImage"
    usedHeapSize : <number>
}
DecodeImage Properties:
{
    children : <object>
    counters : <object>
    data : {
        elementId : <number>
        imageType : "JPEG"
        url : .../inspector/timeline/resources/test.jpg
    }
    endTime : <number>
    startTime : <number>
    type : "DecodeImage"
    usedHeapSize : <number>
}
DecodeImage Properties:
{
    children : <object>
    counters : <object>
    data : {
        elementId : <number>
        imageType : "PNG"
        url : .../inspector/timeline/resources/test.png
    }
    endTime : <number>
    startTime : <number>
    type : "DecodeImage"
    usedHeapSize : <number>
}
DecodeImage Properties:
{
    children : <object>
    counters : <object>
    data : {
        elementId : <number>
        imageType : "WEBP"
        url : .../inspector/timeline/resources/test.webp
    }
    endTime : <number>
    startTime : <number>
    type : "DecodeImage"
    usedHeapSize : <number>
}
DecodeImage Properties:
{
    children : <object>
    counters : <object>
    data : {
        elementId : <number>
        imageType : "PNG"
        url : .../inspector/timeline/resources/big.png
    }
    endTime : <number>
    startTime : <number>
    type : "DecodeImage"
    usedHeapSize : <number>
}
